Add fragment tag and allow escaping opening curly brace
Fragment tag (`<>`) as first non-whitespace string in a line denotes
plaintext (as opposed to Zig code). Allows adding e.g. lines inside a
`<pre>` tag and so on. The `<>` is removed and replaced with whitespace
to ensure the line length/indentation are stable.

Allow escaping literal `{` characters as `\{` to allow using this
character without Zmpl treating it as a data reference opening tag.