Fix issue with templates lacking a terminating linebreak
We already normalize linebreaks to unix format when parsing, but the
parser expects the template to terminate with a linebreak - add this if
it is not already present. Also add a warning requesting a bug report if
the char buf is not empty after reaching the end of the template.