Partial arg type coercion
When passing args to a partial, attempt to coerce a data reference (e.g.
`.foo`, i.e. a `*Value`) to the defined arg type. This allows partials
to receive a literal, local const, etc. as well as a `Value.string`,
`Value.integer`, etc. if the corresponding type is coercable, which
makes partials much more re-usable.