Support modes inside blocks
Allow marking a prefix path as ignored - this allows us to specify a
prefix but to mark it as absent. This is basically a bit of a hack to
let Jetzig run its unit tests when its internal Zmpl templates refer to
templates that are expected to be provided by the application (e.g.
Inertia middleware providing a template that loads a custom header from
the application at a specific path).